extracting abstract class for integration tests with docker
